monorepo: rename 'controller' to 'internal/runtime'
We want to use the top-level dir 'controller' for new controller code
and go module. The existing 'controller' code is a shared set of
listeners for watching COSI resources. Since this code is shared by the
controller and sidecar, it needs to be available to both. It doesn't
need to be made available to other projects, so the Golang 'internal'
dir can be used.
