+++ internal/runtime/interfaces.go
@@ -1,4 +1,4 @@
-package controller
+package runtime
 
 import (
 	"context"
